refactor(shortcuts): 拆分快捷键设置页为 part 结构并抽出公共标签扩展
- shortcut_settings_page.dart 1546 行单文件拆为:页面壳(551) +
  shortcut_settings/action_tile.part.dart(_ActionTile/_MouseChip) +
  shortcut_settings/binding_edit_dialog.part.dart(编辑对话框与鼠标辅助),
  沿用 reader_hibiki 的 part 域拆分范式;公开 API 与 import 路径不变
- 动作/作用域/鼠标绑定标签从页面私有 switch 抽为
  shortcuts/shortcut_labels.dart 公开扩展(action.label/scope.label),
  加动作时标签与数据层就近同步;穷举 switch 保留编译期防漏
- 行为零破坏:注册表/持久化(shortcut_bindings_json schema v5)/写穿路径
  updateBindingWithReassignments→saveShortcutRegistry 均未触碰
- 8 个静态守卫测试同步指向新文件,守卫意图不变(禁词扫描面扩大)
- i18n: 删除死 key shortcut_none(17 语言,i18n_sync --remove + slang 重生成)

flutter analyze 0 issues; flutter test 12243 通过/5 跳过
