+++ hibiki/lib/i18n/strings_ar.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "إعادة تعيين كل الاختصارات في هذا القسم إلى الافتراضي؟",
   "shortcut_keyboard": "لوحة المفاتيح",
   "shortcut_gamepad": "يد التحكّم",
-  "shortcut_none": "بلا",
   "shortcut_clear": "مسح",
   "shortcut_stop_capture": "إيقاف",
   "section_advanced_typography": "متقدم",
