+++ hibiki/lib/i18n/strings_de.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Alle Tastenkürzel in diesem Bereich auf Standard zurücksetzen?",
   "shortcut_keyboard": "Tastatur",
   "shortcut_gamepad": "Gamepad",
-  "shortcut_none": "Keine",
   "shortcut_clear": "Löschen",
   "shortcut_stop_capture": "Stopp",
   "section_advanced_typography": "Erweitert",
