+++ hibiki/lib/i18n/strings_id.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Kembalikan semua pintasan di bagian ini ke default?",
   "shortcut_keyboard": "Keyboard",
   "shortcut_gamepad": "Gamepad",
-  "shortcut_none": "Tidak Ada",
   "shortcut_clear": "Hapus",
   "shortcut_stop_capture": "Berhenti",
   "section_advanced_typography": "Lanjutan",
