+++ hibiki/lib/i18n/strings_it.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Ripristinare tutte le scorciatoie di questa sezione ai valori predefiniti?",
   "shortcut_keyboard": "Tastiera",
   "shortcut_gamepad": "Gamepad",
-  "shortcut_none": "Nessuna",
   "shortcut_clear": "Cancella",
   "shortcut_stop_capture": "Arresta",
   "section_advanced_typography": "Avanzate",
