+++ hibiki/lib/i18n/strings_ja.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "このセクションのすべてのショートカットをデフォルトに戻しますか？",
   "shortcut_keyboard": "キーボード",
   "shortcut_gamepad": "ゲームパッド",
-  "shortcut_none": "なし",
   "shortcut_clear": "クリア",
   "shortcut_stop_capture": "停止",
   "section_advanced_typography": "詳細設定",
