+++ hibiki/lib/i18n/strings_nl.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Alle sneltoetsen in dit gedeelte terugzetten naar standaard?",
   "shortcut_keyboard": "Toetsenbord",
   "shortcut_gamepad": "Gamepad",
-  "shortcut_none": "Geen",
   "shortcut_clear": "Wissen",
   "shortcut_stop_capture": "Stoppen",
   "section_advanced_typography": "Geavanceerd",
