+++ hibiki/lib/i18n/strings_pt-BR.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Restaurar todos os atalhos desta seção para os padrões?",
   "shortcut_keyboard": "Teclado",
   "shortcut_gamepad": "Controle",
-  "shortcut_none": "Nenhum",
   "shortcut_clear": "Limpar",
   "shortcut_stop_capture": "Parar",
   "section_advanced_typography": "Avançado",
