+++ hibiki/lib/i18n/strings_th.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "รีเซ็ตแป้นพิมพ์ลัดทั้งหมดในส่วนนี้เป็นค่าเริ่มต้น?",
   "shortcut_keyboard": "แป้นพิมพ์",
   "shortcut_gamepad": "เกมแพด",
-  "shortcut_none": "ไม่มี",
   "shortcut_clear": "ล้าง",
   "shortcut_stop_capture": "หยุด",
   "section_advanced_typography": "ขั้นสูง",
