+++ hibiki/lib/i18n/strings_tr.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Bu bölümdeki tüm kısayollar varsayılanlara sıfırlansın mı?",
   "shortcut_keyboard": "Klavye",
   "shortcut_gamepad": "Oyun kumandası",
-  "shortcut_none": "Yok",
   "shortcut_clear": "Temizle",
   "shortcut_stop_capture": "Durdur",
   "section_advanced_typography": "Gelişmiş",
