+++ hibiki/lib/i18n/strings_vi.i18n.json
@@ -768,7 +768,6 @@
   "shortcut_reset_confirm": "Khôi phục mọi phím tắt trong mục này về mặc định?",
   "shortcut_keyboard": "Bàn phím",
   "shortcut_gamepad": "Tay cầm",
-  "shortcut_none": "Không có",
   "shortcut_clear": "Xóa",
   "shortcut_stop_capture": "Dừng",
   "section_advanced_typography": "Nâng cao",
